Content Updates: 01/08/2022 to 01/16/2022 (#15)
* feat: updateContent.sh updated to allow arg parsing

* chore: content update, Sat Jan  8 00:16:37 PST 2022

* chore: content update, Sat Jan  8 00:36:52 PST 2022

* chore: content update, Sat Jan  8 01:19:45 PST 2022

* chore: content update, Sat Jan  8 03:10:03 PST 2022

* chore: content update, Sat Jan  8 17:10:03 PST 2022

* chore: content update, Sun Jan  9 03:09:58 PST 2022

* chore: content update, Sun Jan  9 17:10:22 PST 2022

* chore: content update, Mon Jan 10 03:10:00 PST 2022

* chore: content update, Mon Jan 10 17:10:04 PST 2022

* chore: content update, Tue Jan 11 03:09:59 PST 2022

* chore: content update, Tue Jan 11 17:10:02 PST 2022

* chore: content update, Wed Jan 12 03:09:57 PST 2022

* chore: content update, Wed Jan 12 17:09:59 PST 2022

* chore: content update, Thu Jan 13 03:09:53 PST 2022

* chore: content update, Thu Jan 13 17:10:02 PST 2022

* chore: content update, Fri Jan 14 03:10:03 PST 2022

* chore: content update, Fri Jan 14 17:10:05 PST 2022

* chore: content update, Sat Jan 15 03:10:04 PST 2022

* chore: content update, Sat Jan 15 17:10:03 PST 2022

* chore: content update, Sun Jan 16 03:09:55 PST 2022

+++ src/tools/updateContent.sh
@@ -8,26 +8,48 @@ CONFIG_FILES=("bktParams.js" "coursePlans.js" "skillModel.js")
 
 OUT_FILE="$CURRENT_PATH/logs/updateContent-$(date -d "today" +"%Y%m%d%H%M").log"
 
+DO_UPDATE=true
+
+while [[ "$#" -gt 0 ]]; do
+    case $1 in
+    -c | --content)
+        CONTENT_DIR_NAME="$2"
+        shift
+        ;;
+    -o | --out)
+        OUT_FILE="$2"
+        shift
+        ;;
+    -n | --no-update) DO_UPDATE=false ;;
+    *)
+        echo "Unknown parameter passed: $1"
+        exit 1
+        ;;
+    esac
+    shift
+done
+
 cd "../../../$CONTENT_DIR_NAME" || (
-  echo "Could not access content directory: $CONTENT_DIR_NAME, exiting early.."
-  exit
+    echo "Could not access content directory: $CONTENT_DIR_NAME, exiting early.."
+    exit
 )
 
 CONTENT_PATH=$(pwd)
 
 if [ ! -d "$SUBDIRECTORY" ]; then
-  echo "$SUBDIRECTORY sub directory does not exist."
-  exit
+    echo "$SUBDIRECTORY sub directory does not exist."
+    exit
 fi
 
-
-git remote update
-git reset --hard origin/main
-git pull
+if [[ "$DO_UPDATE" = true ]]; then
+    git remote update
+    git reset --hard origin/main
+    git pull
+fi
 
 if [ ! -d "$SUBDIRECTORY" ]; then
-  echo "$SUBDIRECTORY sub directory does not exist."
-  exit
+    echo "$SUBDIRECTORY sub directory does not exist."
+    exit
 fi
 
 cd "$CURRENT_PATH" || exit
@@ -48,7 +70,7 @@ echo "Copying over new Config Files"
 cd "$CONTENT_PATH" || exit
 
 for FILE in "${CONFIG_FILES[@]}"; do
-  cp "$FILE" "$ROOT_PATH/src/config"
+    cp "$FILE" "$ROOT_PATH/src/config"
 done
 
 cd "$ROOT_PATH" || exit
@@ -61,7 +83,8 @@ cp src/config/bktParams/bktParams1.js src/config/bktParams/bktParams2.js
 
 echo "Generating flattened problem pool."
 
-cd src/tools
-node generateFlatProblemPool.js >"$OUT_FILE" 2>&1
+cd src/tools || exit
+node generateFlatProblemPool.js >>"$OUT_FILE" 2>&1
 
 echo "Make sure to increment version in OpenITS/src/config/config.js!"
+
